requirements.txt
google-generativeai
google-genai
chromadb
requests
chardet
PyPDF2
python-docx
